fix: [RUSAA-40] split routing tests to pass 600-line file-size cap
Move 8 query-routing integration tests out of integration.rs (611 lines)
into a new tests/routing.rs (304 lines). integration.rs returns to 329
lines. Both files are well under the 600-line hard cap.

CI tenant-isolation job updated to run --test integration --test routing.
